♻️ Refacor Declaration Syntax
refactor: improve declaration syntax and syntax error message
