--- examples/integer.ll
@@ -1,3 +1,0 @@
-((100-10)/2);
-5 - - - + - (3 + 4) - +2;
-((-100)-10)/+2;
