--- examples/string.ll
@@ -1,1 +1,0 @@
->> "\tHello\nEnd";
