--- examples/visitor.ll
@@ -1,7 +1,0 @@
-var a -> number := 2;
->> a;
-
-var b -> () := {
-  >> "Hello";
-}
-b();
