--- examples/while.ll
@@ -1,6 +1,0 @@
-NUMBER cnt = 4;
-
-WHILE ((10 < cnt) < 1 && cnt < 10 && 1 < 2) {
-  >> "Cnt:" >> cnt;
-  cnt = cnt + 2;
-}
